Moved Search page to OSF components.
