+++ lib/osf-components/addon/components/search-page/styles.scss
