+++ lib/osf-components/app/components/search-page/component.js
@@ -1,0 +1,1 @@
+export { default } from 'osf-components/components/search-page/component';
